+++ packages/eslint-config/react.js
@@ -1,6 +1,0 @@
-const { resolveModule } = require('./src/utils');
-
-module.exports = {
-  env: { browser: true },
-  extends: ['./index', ...['react'].map(resolveModule)],
-};