+++ packages/eslint-config/src/rules/react/emotion.js
@@ -1,0 +1,10 @@
+module.exports = {
+  plugins: ['@emotion'],
+  rules: {
+    '@emotion/import-from-emotion': 'error',
+    '@emotion/jsx-import': 'error',
+    '@emotion/no-vanilla': 'error',
+    '@emotion/pkg-renaming': 'error',
+    '@emotion/styled-import': 'error',
+  },
+};