+++ packages/eslint-config/src/utils.js
@@ -1,5 +1,7 @@
+const path = require('path');
+
 function resolveModule(moduleName) {
-  return require.resolve(`./rules/${moduleName}`);
+  return require.resolve(path.resolve(process.cwd(), `./src/rules/${moduleName}`));
 }
 
 module.exports = {